Refactor test helpers in own file
