+++ runner/lib/test-local-chain.js
@@ -1,117 +1,25 @@
 /* global process Buffer */
 
-import chalk from 'chalk';
-
 import { dirname } from 'path';
 import { promisify } from 'util';
 import { PassThrough, pipeline as pipelineCallback } from 'stream';
-
-// TODO: pass an "httpRequest" as power instead of importing
-import http from 'http';
 
 import {
   childProcessDone,
   makeSpawnWithPrintAndPipeOutput,
 } from './helpers/child-process.js';
 import LineStreamTransform from './helpers/line-stream-transform.js';
-import { PromiseAllOrErrors, sleep, tryTimeout } from './helpers/async.js';
-import { makeOutputter } from './helpers/outputter.js';
+import { PromiseAllOrErrors, tryTimeout } from './helpers/async.js';
 import { whenStreamSteps } from './helpers/stream-steps.js';
+import {
+  getArgvMatcher,
+  getChildMatchingArgv,
+  wrapArgvMatcherIgnoreEnvShebang,
+  httpRequest,
+  getConsoleAndStdio,
+} from './test-helpers.js';
 
 const pipeline = promisify(pipelineCallback);
-
-/**
- * @param {string | URL} url
- * @param {http.RequestOptions & {body?: Buffer}} options
- * @returns {Promise<http.IncomingMessage>}
- */
-const httpRequest = (url, options) => {
-  return new Promise((resolve, reject) => {
-    const { body, ...httpOptions } = options;
-
-    const req = http.request(url, httpOptions);
-    req.on('response', resolve).on('error', reject);
-    if (body) {
-      req.write(body);
-    }
-    req.end();
-  });
-};
-
-/**
- *
- * @param {import('./helpers/process-info.js').ProcessInfo} info
- * @param {number} [retries]
- * @returns {Promise<string[] | null>}
- */
-const untilArgv = async (info, retries = 50) => {
-  const argv = await info.getArgv();
-  return (
-    argv ||
-    (retries > 0 ? (await sleep(100), untilArgv(info, retries - 1)) : null)
-  );
-};
-
-/**
- *
- * @param {import('./helpers/process-info.js').ProcessInfo} info
- * @param {number} [retries]
- * @returns {Promise<import('./helpers/process-info.js').ProcessInfo[]>}
- */
-const untilChildren = async (info, retries = 50) => {
-  const children = await info.getChildren();
-  return children.length || retries === 0
-    ? children
-    : (await sleep(100), untilChildren(info, retries - 1));
-};
-
-/** @typedef {(argv: string[]) => boolean} ArgvMatcher */
-
-/**
- * @param {(RegExp | null | undefined)[]} argMatchers
- * @returns {ArgvMatcher}
- */
-const getArgvMatcher = (argMatchers) => (argv) =>
-  argv.every((arg, idx) => {
-    const matcher = argMatchers[idx];
-    return !matcher || matcher.test(arg);
-  });
-
-/**
- * @param {ArgvMatcher} argvMatcher
- * @returns {ArgvMatcher}
- */
-const wrapArgvMatcherIgnoreEnvShebang = (argvMatcher) => (argv) =>
-  argvMatcher(argv) || (/env$/.test(argv[0]) && argvMatcher(argv.slice(1)));
-
-/**
- * @param {import('./helpers/process-info.js').ProcessInfo} launcherInfo
- * @param {ArgvMatcher} argvMatcher
- */
-const getChildMatchingArgv = async (launcherInfo, argvMatcher) => {
-  const childrenWithArgv = await Promise.all(
-    (await untilChildren(launcherInfo)).map(async (info) => ({
-      info,
-      argv: await untilArgv(info),
-    })),
-  );
-
-  const result = childrenWithArgv.find(({ argv }) => argv && argvMatcher(argv));
-
-  if (result) {
-    return result.info;
-  }
-
-  console.error(
-    `getChildMatchingArgv: ${
-      childrenWithArgv.length
-    } child process, none of ["${childrenWithArgv
-      .map(({ argv }) => (argv || ['no argv']).join(' '))
-      .join('", "')}"] match expected arguments`,
-  );
-
-  throw new Error("Couldn't find child process");
-};
 
 const chainDirPrefix = '_agstate/agoric-servers/local-chain-';
 
@@ -153,22 +61,6 @@ export const makeTestOperations = ({
     spawn,
     end: false,
   });
-
-  /**
-   * @param {string} prefix
-   * @param {import("stream").Writable} stdout
-   * @param {import("stream").Writable} stderr
-   * @returns {{stdio: [undefined, import("stream").Writable, import("stream").Writable], console: Console}}
-   */
-  const getConsoleAndStdio = (prefix, stdout, stderr) => {
-    const { console, out, err } = makeOutputter({
-      out: stdout,
-      err: stderr,
-      outPrefix: prefix && `${chalk.bold.blue(prefix)}: `,
-      errPrefix: prefix && `${chalk.bold.red(prefix)}: `,
-    });
-    return { console, stdio: [undefined, out, err] };
-  };
 
   return harden({
     setupTest: async ({ stdout, stderr, config = {} }) => {
